+++ src/main/java/kr/co/pawong/pwbe/adoption/application/port/in/ApiAdoptionUseCase.java
@@ -1,6 +1,6 @@
 package kr.co.pawong.pwbe.adoption.application.port.in;
 
-import kr.co.pawong.pwbe.adoption.application.port.in.dto.AdoptionCreate;
+import kr.co.pawong.pwbe.infrastructure.api.dto.AdoptionCreate;
 import kr.co.pawong.pwbe.adoption.application.service.dto.AdoptionApi;
 
 public interface ApiAdoptionUseCase {
