+++ src/main/java/kr/co/pawong/pwbe/adoption/application/port/in/CommandAdoptionDataUseCase.java
@@ -1,7 +1,7 @@
 package kr.co.pawong.pwbe.adoption.application.port.in;
 
 import java.util.List;
-import kr.co.pawong.pwbe.adoption.application.port.in.dto.AdoptionCreate;
+import kr.co.pawong.pwbe.infrastructure.api.dto.AdoptionCreate;
 import kr.co.pawong.pwbe.adoption.domain.model.Adoption;
 
 public interface CommandAdoptionDataUseCase {
