+++ src/main/java/kr/co/pawong/pwbe/adoption/domain/model/Adoption.java
@@ -2,7 +2,7 @@ package kr.co.pawong.pwbe.adoption.domain.model;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import kr.co.pawong.pwbe.adoption.application.port.in.dto.AdoptionCreate;
+import kr.co.pawong.pwbe.infrastructure.api.dto.AdoptionCreate;
 import kr.co.pawong.pwbe.adoption.application.port.in.dto.AdoptionUpdate;
 import kr.co.pawong.pwbe.adoption.enums.ActiveState;
 import kr.co.pawong.pwbe.adoption.enums.NeuterYn;
