+++ src/main/java/kr/co/pawong/pwbe/shelter/adapter/in/api/ShelterUpdateController.java
@@ -1,6 +1,6 @@
 package kr.co.pawong.pwbe.shelter.adapter.in.api;
 
-import kr.co.pawong.pwbe.shelter.application.port.in.dto.ShelterCreate;
+import kr.co.pawong.pwbe.infrastructure.api.dto.ShelterCreate;
 import kr.co.pawong.pwbe.infrastructure.api.ApiShelterService;
 import kr.co.pawong.pwbe.shelter.application.port.in.UpdateShelterDataUseCase;
 import lombok.RequiredArgsConstructor;
