+++ src/main/java/kr/co/pawong/pwbe/shelter/application/port/in/UpdateShelterDataUseCase.java
@@ -2,7 +2,7 @@ package kr.co.pawong.pwbe.shelter.application.port.in;
 
 import java.util.List;
 import kr.co.pawong.pwbe.adoption.application.port.out.dto.AdoptionCareDto;
-import kr.co.pawong.pwbe.shelter.application.port.in.dto.ShelterCreate;
+import kr.co.pawong.pwbe.infrastructure.api.dto.ShelterCreate;
 
 
 public interface UpdateShelterDataUseCase {
